JSK V_3 fix signup bug

+++ backend/Controller/Authentication/signup.js
@@ -21,13 +21,15 @@ const signUp = async (req, res) => {
       $or: [{ email: email }, { phoneNumber: phoneNumber }],
     });
 
-    if (isAlreadyUser.verified==true) {
+    
+
+    if (isAlreadyUser && isAlreadyUser.verified==true) {
       return res
         .status(400)
         .json({ success: false, message: "User already exists."});
     }
 
-    if(isAlreadyUser.verified==false){
+    if(isAlreadyUser && isAlreadyUser.verified==false){
       await User.deleteOne({ _id: isAlreadyUser._id });
     }
